# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XLS-66 Lending Protocol Reference App
  description: |
    HTTP API for the XRPL Lending Protocol (XLS-66) + Single Asset Vault
    (XLS-65) reference implementation. Every route that mutates on-chain
    state maps 1:1 to an XRPL transaction and re-queries the ledger before
    returning.

    ## Conventions

    - **Auth** — handled by Auth0 Universal Login. Clients obtain a session
      by visiting `/auth/login` (browser only) — there is no programmatic
      login endpoint. The session is stored in an httpOnly cookie `xls66-auth`.
      Every `/api/*` route is gated by `src/middleware.ts` and returns
      `401 Unauthorized` without a valid cookie.
    - **Amounts** — drops (integer string) for XRP; decimal string for IOU
      and MPT. See the README for the scaling conventions.
    - **Rates** — UI-facing rates are % or basis points; on-ledger they are
      1/10 bps. The API accepts the UI form and scales at the boundary.
    - **Content type** — `application/json` for all request/response bodies.
  version: 1.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: XLS-66 reference app
servers:
  - url: http://localhost:3000
    description: Local dev server
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xls66-auth
      description: httpOnly cookie set by `POST /api/session`. Value is the Session document's MongoDB ObjectId.
...
  schemas:
...
    TxResult:
      description: Pass-through of the rippled `tx` submission result. Only the hash is canonical; everything else is for debugging.
      type: object
      additionalProperties: true
      properties:
        hash: { type: string }
        ledger_index: { type: integer }
        meta:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid `xls66-auth` cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/vault/withdraw:
    post:
      tags: [Vault]
      summary: VaultWithdraw
      description: Redeem shares for underlying asset. Payload mirrors `/api/vault/deposit`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vaultId]
              properties:
                vaultId: { type: string }
                amountDrops: { type: string }
                tokenAmount: { type: string }
      responses:
        "200":
          description: Withdraw submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...